Refactoring ZMQ Thread
